Esign
Esign failed when eye forms were used during an encounter.  The structure of OpenEMR requires the name of the form to be the samepart of the table name (form_eye_mag).  For the Eye Form the name of the form is eye_mag but the name of the table is form_eye_base.